Revert "Revert "DEVPROD-12862: Remove the performance endpoints on Cedar (#55…"
This reverts commit 38509dbe5ae9d1d0fdfcec525440ed9807edb019.
