+++ rest/client.go
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	"github.com/evergreen-ci/cedar"
-	"github.com/evergreen-ci/cedar/rest/model"
 	"github.com/evergreen-ci/gimlet"
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/message"
@@ -587,70 +586,4 @@ func (c *Client) GetUserCertificate(ctx context.Context, username, password, api
 
 func (c *Client) GetUserCertificateKey(ctx context.Context, username, password, apiKey string) (string, error) {
 	return c.authCredRequest(ctx, c.getURL("/v1/admin/users/certificate/key"), username, password, apiKey)
-}
-
-func (c *Client) FindPerformanceResultById(ctx context.Context, id string) (*model.APIPerformanceResult, error) {
-	url := c.getURL(fmt.Sprintf("/v1/perf/%s", id))
-
-	req, err := c.makeRequest(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "building request")
-	}
-
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, errors.Wrap(err, "making request")
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		srverr := gimlet.ErrorResponse{}
-		if err = gimlet.GetJSON(resp.Body, &srverr); err != nil {
-			return nil, errors.Wrap(err, "parsing error message")
-		}
-
-		return nil, srverr
-	}
-
-	out, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, errors.Wrap(err, "reading response")
-	}
-	result := &model.APIPerformanceResult{}
-	if err = json.Unmarshal(out, result); err != nil {
-		return nil, errors.Wrap(err, "unmarshalling response data")
-	}
-
-	return result, nil
-}
-
-func (c *Client) RemovePerformanceResultById(ctx context.Context, id string) (string, error) {
-	url := c.getURL(fmt.Sprintf("/v1/perf/%s", id))
-
-	req, err := c.makeRequest(ctx, http.MethodDelete, url, nil)
-	if err != nil {
-		return "", errors.Wrap(err, "building request")
-	}
-
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return "", errors.Wrap(err, "making request")
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		srverr := gimlet.ErrorResponse{}
-		if err = gimlet.GetJSON(resp.Body, &srverr); err != nil {
-			return "", errors.Wrap(err, "parsing error message")
-		}
-
-		return "", srverr
-	}
-
-	out, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", errors.Wrap(err, "reading response")
-	}
-
-	return string(out), nil
 }
