+++ rest/model/buildlogger_test.go
@@ -12,7 +12,7 @@ import (
 func TestBuildloggerImport(t *testing.T) {
 	t.Run("InvalidType", func(t *testing.T) {
 		apiLog := &APILog{}
-		assert.Error(t, apiLog.Import(dbmodel.PerformanceResult{}))
+		assert.Error(t, apiLog.Import(dbmodel.TestLogInfo{}))
 	})
 	t.Run("ValidLog", func(t *testing.T) {
 		log := dbmodel.Log{
