+++ rpc/internal/buildlogger_service_test.go
@@ -21,6 +21,10 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
+const (
+	testDBName = "cedar_grpc_test"
+)
+
 func TestCreateLog(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
